organize project; make static value to use in place of credential package
