+++ configs/tsconfig.eslint.json
@@ -5,6 +5,7 @@
   },
   "include": [
     "../src/**/*.ts",
-    "../tests/**/*.ts"
+    "../tests/**/*.ts",
+    "../lib/**/*.ts"
   ]
 }
