+++ package.json
@@ -51,7 +51,6 @@
     "echo-cli": "^2.0.0",
     "jest": "^29.7.0",
     "jest-ci-spec-reporter": "^1.0.3",
-    "move-file-cli": "^3.0.0",
     "ts-jest": "^29.1.2",
     "ts-node": "^10.9.2",
     "ts-standard": "^12.0.2",
@@ -65,5 +64,8 @@
       "dist"
     ],
     "project": "./configs/tsconfig.eslint.json"
+  },
+  "dependencies": {
+    "jsonwebtoken": "^9.0.2"
   }
 }
